requirements.txt
pyyaml
dependency-injector
